+++ demisto_sdk/common/scripts/dev_scripts/run_dev_tasks.sh
@@ -1,9 +1,9 @@
 #!/bin/sh
 
-# Run pylint and pytest in the current directory. 
-# Used by pkg_dev_test_tasks.py to run pylint and pytest 
+# Run pylint and pytest in the current directory.
+# Used by pkg_dev_test_tasks.py to run pylint and pytest
 # inside a docker. Since this is meant to run inside a minimal docker
-# image it uses sh and not bash. Additionally, script tries to keep it 
+# image it uses sh and not bash. Additionally, script tries to keep it
 # simply and not use any shell utilities that may be missing in a minimal docker.
 
 # Env variables:
@@ -15,7 +15,7 @@
 
 pylint_return=0
 if [ -z "${PYLINT_SKIP}" ]; then
-    echo "======== Running pylint on files: ${PYLINT_FILES} ==========="
+    echo "=============== Running pylint on files: ${PYLINT_FILES} ==============="
     python -m pylint -E -e string -d duplicate-string-formatting-argument -f parseable --generated-members=requests.packages.urllib3,requests.codes.ok \
         ${PYLINT_FILES}
     pylint_return=$?
@@ -23,6 +23,7 @@ if [ -z "${PYLINT_SKIP}" ]; then
 fi
 
 if [ -z "${PYTEST_SKIP}" ]; then
+    echo ""
     echo "========= Running pytest ==============="
 fi
 
@@ -38,7 +39,7 @@ if [ -z "${PYTEST_SKIP}" -a -z "${PYTEST_FAIL_NO_TESTS}" ]; then
             echo "========= No tests found. Skipping. ========"
             echo "========= Output of: pytest --collect-only ========"
             echo "$collect_res"
-            echo "========= End of Output ========"
+            echo -e "\e[32m=============== End of Output for: ${PYLINT_FILES} ===============\e[39m"
             PYTEST_SKIP=1
         ;;
     esac
@@ -55,5 +56,5 @@ if [ $pylint_return -ne 0 -o $pytest_return -ne 0 ]; then
     echo "=========== ERRORS FOUND ===========" 1>&2
     echo "pylint/pytest returned errors. pylint: [$pylint_return], pytest: [$pytest_return]" 1>&2
     echo "====================================" 1>&2
-    exit 3 
+    exit 3
 fi