+++ CHANGELOG.md
@@ -4,6 +4,29 @@
 
 [1]: https://pypi.org/project/demisto-sdk/#history
 
+### 0.2.7
+
+* Added support for multi-package **lint** both with parallel and without.
+* Added all parameter in **lint** to run on all packages and packs in content repository.
+
+### 0.2.6
+
+* Fixed an issue with locating release notes for beta integrations in **validate**.
+
+### 0.2.5
+
+* Fixed an issue with locating release notes for beta integrations in **validate**.
+
+### 0.2.4
+
+* Adding image validation to Beta_Integration and Packs in **validate**.
+
+### 0.2.3
+
+* Adding Beta_Integration to the structure validation process.
+* Fixing bug where **validate** did checks on TestPlaybooks.
+* Added requirements parameter to **lint**.
+
 ### 0.2.2
 
 * Fixing bug where **lint** did not return exit code 1 on failure.